Add new preference - enabled
Add a new user preference - enabled - which allows the extension to be temporarily turned off. This can be useful if you suspect the extension is causing a problem with Brackets or another extension. The default for this setting is 'true' (the extension is enabled).
Make minor changes to the README.md for the new setting and also to expand on the 'Install' section.